Move outputs to api (currently includes testing Outputs) and lsl_output
