+++ src/neural_data_simulator/core/encoder.py
@@ -5,7 +5,7 @@ from typing import Iterator, Optional, Protocol, runtime_checkable
 
 from neural_data_simulator.core.inputs.api import Input
 from neural_data_simulator.core.models import EncoderModel
-from neural_data_simulator.core.outputs import Output
+from neural_data_simulator.core.outputs.api import Output
 from neural_data_simulator.core.samples import Samples
 from neural_data_simulator.util.runtime import open_connection
