+++ src/neural_data_simulator/decoder/run_decoder.py
@@ -7,10 +7,11 @@ from typing import cast
 
 from pydantic_yaml import VersionedYamlModel
 
-from neural_data_simulator.core import outputs
 from neural_data_simulator.core import timing
 from neural_data_simulator.core.inputs import api as inputs
 from neural_data_simulator.core.inputs.lsl_input import LSLInput
+from neural_data_simulator.core.outputs import api as outputs
+from neural_data_simulator.core.outputs.lsl_output import LSLOutputDevice
 from neural_data_simulator.core.settings import LogLevel
 from neural_data_simulator.core.settings import TimerModel
 from neural_data_simulator.decoder.decoders import Decoder
@@ -81,7 +82,7 @@ def run():
 
     # Create LSL input and output objects
     output_settings = settings.decoder.output
-    data_output = outputs.LSLOutputDevice.from_lsl_settings(
+    data_output = LSLOutputDevice.from_lsl_settings(
         lsl_settings=output_settings.lsl,
         sampling_rate=output_settings.sampling_rate,
         n_channels=output_settings.n_channels,
