+++ src/neural_data_simulator/scripts/run_ephys_generator.py
@@ -17,7 +17,6 @@ from typing import cast, Optional
 
 import numpy as np
 
-from neural_data_simulator.core import outputs
 from neural_data_simulator.core.ephys_generator import ContinuousData
 from neural_data_simulator.core.ephys_generator import NoiseData
 from neural_data_simulator.core.ephys_generator import ProcessOutput
@@ -28,7 +27,8 @@ from neural_data_simulator.core.inputs.api import SpikeRateInput
 from neural_data_simulator.core.inputs.lsl_input import LSLInput
 from neural_data_simulator.core.inputs.lsl_input import LSLSpikeRateInputAdapter
 from neural_data_simulator.core.inputs.testing_input import SpikeRateTestingInput
-from neural_data_simulator.core.outputs import StreamConfig
+from neural_data_simulator.core.outputs.lsl_output import LSLOutputDevice
+from neural_data_simulator.core.outputs.lsl_output import StreamConfig
 from neural_data_simulator.core.settings import EphysGeneratorEndpointType
 from neural_data_simulator.core.settings import EphysGeneratorSettings
 from neural_data_simulator.core.settings import Settings
@@ -77,7 +77,7 @@ def _setup_LSL_input(
     return spike_rate_input
 
 
-def _setup_LSL_output(config: StreamConfig) -> outputs.LSLOutputDevice:
+def _setup_LSL_output(config: StreamConfig) -> LSLOutputDevice:
     """Set up output that will make the data available via an LSL stream.
 
     Args:
@@ -87,7 +87,7 @@ def _setup_LSL_output(config: StreamConfig) -> outputs.LSLOutputDevice:
         An LSL output stream that can be used by the ephys generator
           to publish data.
     """
-    lsl_output = outputs.LSLOutputDevice(config)
+    lsl_output = LSLOutputDevice(config)
     lsl_output.connect()
     return lsl_output
