+++ tests/test_encoder.py
@@ -5,7 +5,7 @@ import numpy as np
 
 import neural_data_simulator.core.encoder as encoder
 from neural_data_simulator.core.inputs import api as inputs
-import neural_data_simulator.core.outputs as outputs
+import neural_data_simulator.core.outputs.api as outputs
 from neural_data_simulator.core.samples import Samples
